old_password refactoring
