+++ lib/devise-security/models/active_record/old_password.rb
@@ -1,0 +1,6 @@
+# frozen_string_literal: true
+
+require 'active_record'
+class OldPassword < ActiveRecord::Base
+  belongs_to :password_archivable, polymorphic: true
+end
